start moving store impl into iroh-bytes
