+++ iroh-bytes/src/lib.rs
@@ -8,9 +8,10 @@ pub mod get;
 pub mod hashseq;
 pub mod protocol;
 pub mod provider;
+pub mod store;
 pub mod util;
 
-pub use crate::util::Hash;
+pub use crate::util::{Hash, Tag, TempTag};
 use bao_tree::BlockSize;
 
 /// Block size used by iroh, 2^4*1024 = 16KiB
