+++ iroh-bytes/src/store.rs
@@ -1,0 +1,2 @@
+//! Implementations of blob stores
+pub mod readonly_mem;
