+++ iroh/src/baomap/mem.rs
@@ -33,10 +33,8 @@ use iroh_bytes::baomap::EntryStatus;
 use iroh_bytes::baomap::ExportMode;
 use iroh_bytes::baomap::ImportMode;
 use iroh_bytes::baomap::ImportProgress;
-use iroh_bytes::baomap::LivenessTracker;
 use iroh_bytes::baomap::PartialMap;
 use iroh_bytes::baomap::PartialMapEntry;
-use iroh_bytes::baomap::TempTag;
 use iroh_bytes::baomap::ValidateProgress;
 use iroh_bytes::baomap::{Map, MapEntry, ReadableStore};
 use iroh_bytes::util::progress::IdGenerator;
@@ -45,7 +43,9 @@ use iroh_bytes::util::progress::ProgressSender;
 use iroh_bytes::util::runtime;
 use iroh_bytes::util::BlobFormat;
 use iroh_bytes::util::HashAndFormat;
+use iroh_bytes::util::LivenessTracker;
 use iroh_bytes::util::Tag;
+use iroh_bytes::TempTag;
 use iroh_bytes::{Hash, IROH_BLOCK_SIZE};
 use iroh_io::AsyncSliceReader;
 use iroh_io::AsyncSliceWriter;
